Extract shared comment and reaction group code
